#24 Refactoring titles
