+++ dev/App.css
@@ -5,3 +5,7 @@
 label {
   padding-top: 15px;
 }
+
+body {
+    background-color: #ccc;
+}
