+++ dev/actions/actionCreators.js
@@ -2,7 +2,7 @@ import fetch from 'isomorphic-fetch';
 import ArticlesApi from '../api/ArticlesApi';
 
 export const ADD_ARTICLE = 'ADD_ARTICLE';
-export const ADD_ALL_ARTICLE = 'ADD_ALL_ARTICLE';
+export const GET_ALL_ARTICLES = 'GET_ALL_ARTICLES';
 
 //add Article
 export function addArticle(article) {
@@ -20,7 +20,7 @@ export function fetchAllArticles() {
     return function(dispatch) {
         return ArticlesApi.getAllArticles().then( response => {
             dispatch({
-                type: ADD_ALL_ARTICLE,
+                type: GET_ALL_ARTICLES,
                 payload: response
             });
         }).catch(error => {
